New GMP-Starter, New Functions and fixes

+++ GMP_Server/Network/Messages/Connection/ConnectionMessage.cs
@@ -49,7 +49,7 @@ namespace GUC.Server.Network.Messages.Connection
 
             stream.Write(Player.sSendAllKeys);
             stream.Write(Player.sSendKeys.Count);
-            foreach (int k in Player.sSendKeys)
+            foreach (byte k in Player.sSendKeys)
             {
                 stream.Write(k);
             }
